+++ lib/openmhealth_schemas.dart
@@ -1,8 +1,9 @@
 library openmhealth_schemas;
 
-export 'package:openmhealth_schemas/domain/omh/schema_support.dart';
+export 'package:openmhealth_schemas/domain/omh/schemas.dart';
 export 'package:openmhealth_schemas/domain/omh/times.dart';
 export 'package:openmhealth_schemas/domain/omh/measure.dart';
 export 'package:openmhealth_schemas/domain/omh/units.dart';
 export 'package:openmhealth_schemas/domain/omh/statistics.dart';
 export 'package:openmhealth_schemas/domain/omh/geoposition.dart';
+export 'package:openmhealth_schemas/domain/omh/activity.dart';